# sso-service/docs/swagger.yaml
basePath: /
definitions:
...
  response.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      metadata: {}
      status:
        type: string
      status_code:
        type: integer
    type: object
  school_dto.CreateSchoolRequestDTO:
    properties:
      name:
        type: string
      region:
        type: string
    type: object
  school_dto.SchoolResponseDTO:
    properties:
      id:
        type: string
      name:
        type: string
      region:
        type: integer
    type: object
  school_dto.UpdateSchoolRequestDTO:
    properties:
      name:
        type: string
      region:
        type: string
    type: object
  user_dto.SearchAttributesUserDTO:
    properties:
      activated:
        type: boolean
      birthdate:
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: integer
      id:
        type: string
      patronymic:
        type: string
      phone_number:
        type: string
      role:
        type: integer
      surname:
        type: string
    type: object
  user_dto.UpdateUserRequestDTO:
    properties:
      birthdate:
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: string
      password:
        type: string
      patronymic:
        type: string
      phone_number:
        type: string
      surname:
        type: string
    type: object
  user_dto.UserParticipantResponseDTO:
    properties:
      activated:
        type: boolean
      birthdate:
        type: string
      citizenship:
        type: string
      class_number:
        type: string
      disability:
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: integer
      id:
        type: string
      patronymic:
        type: string
      phone_number:
        type: string
      role:
        type: integer
      school_id:
        type: string
      surname:
        type: string
      user_id:
        type: string
    type: object
  user_dto.UserResponseDTO:
    properties:
      activated:
        type: boolean
      birthdate:
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: integer
      id:
        type: string
      patronymic:
        type: string
      phone_number:
        type: string
      role:
        type: integer
      surname:
        type: string
    type: object
host: 172.16.1.39:8181
info:
  contact: {}
  description: Документация к микросервису авторизации
  title: SSO Service API
  version: "1.0"
paths:
...
  /api/schools:
    get:
      description: Получение всех школ
      parameters:
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Ограничение на количество записей
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/school_dto.SchoolResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
    post:
      consumes:
      - application/json
      description: Создание школы
      parameters:
      - description: Данные для создания школы
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/school_dto.CreateSchoolRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
  /api/schools/{id}:
    get:
      description: Получение школы по id
      parameters:
      - description: id школы
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/school_dto.SchoolResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
    put:
      consumes:
      - application/json
      description: Обновление школы
      parameters:
      - description: Данные для обновления школы
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/school_dto.UpdateSchoolRequestDTO'
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
  /api/schools/count:
    get:
      description: Получение количества школ
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
  /api/users:
    get:
      description: Получение всех пользователей
      parameters:
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Ограничение на количество записей
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/user_dto.UserResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuths: []
      tags:
      - users
  /api/users/{id}:
    delete:
      description: Удаление пользователя
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - users
    get:
      description: Получение пользователя по id
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/user_dto.UserResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновление пользователя
      parameters:
      - description: Данные для обновления пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user_dto.UpdateUserRequestDTO'
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - users
  /api/users/all-info/{id}:
    get:
      description: Получение всей информации о пользователе ученике
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/user_dto.UserParticipantResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - users
  /api/users/count:
    get:
      description: Получение количесва пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - users
  /api/users/filter:
    post:
      consumes:
      - application/json
      description: Получение пользователя по фильтру из его полей
      parameters:
      - description: Данные для поиска пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user_dto.SearchAttributesUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/user_dto.UserResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuths: []
      tags:
      - users
  /api/users/list:
    post:
      consumes:
      - application/json
      description: Получение пользователей по списку id
      parameters:
      - description: Список id пользователей
        in: body
        name: credentials
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/user_dto.UserResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"